Fix issue selecting specific BSQ UTX for send.
The BisqDefaultCoinSelector is invoked multiple times when a
transaction is being built.  Problem was that the coin selection
was being reset within, ultimately leading to coin selection
being completely ignored.  Solution is to reset the selection
after the transaction has been built.

The UI for selecting UTXO inputs was being resized to the number
of UTXO available, which caused problems when there is a large
UTXO set.  Solution is to size the UI to accomodate 3 to 15 rows,
which provides a reasonable screen display and a scrollbar is
provided for the case of a large UTXO set.