Add jitpack.yml with before_install instruction to run 'git lfs pull'
Jitpack's current default version of git-lfs works, but this yaml file
may need to download, install, and specify jitpack build's LFS version
in the future.

By default, jitpack is able to build bisq with OpenJDK 11.0.2, but
this yaml file may be to download, install, and specify jitpack
build's JDK in the future (if/when jitpack doesn't have the required
JDK).