Fix Total BSQ Supply and Change of BSQ Supply calculations.
Total BSQ Supply needs to be cumulative from genesis to datapoint.
Change of BSQ Supply needs to handle situations where there is no
issuange to offset a burn.  There was a bug in the use of
getMergedMap(a - b) which would result in a burn being counted as
an issuance in that case.  Solved by making burns negative and using
Long::sum instead of a - b.