Use recovery for all RPC HttpExceptions exceptions in generic error handler.
It seems that a reorgs or if a new block arrives quickly after another, that bitcoind is not ready
for a request and throws an RPC error, also the blockhandler throws that if the new block came too fast.
We try to request again with increasing delays up to 1 min. If after 10 attempts its still failing we
do a reorg and start parting again.