Fix memory leak in PriceRequest thread name.
The thread name is being appended at each price request with the
request URL.  This is a leak because the size of the string forever
increases.

There's actually no benefit to modifying the thread name with
the request URL, as whatever small amount of logging is done in
the price subsystem includes the applicable url in the message
itself.  So applying the KISS rule, we no longer munge the
thread name at all.